Update popup.js
Changed the scrapeContentFromPage function to account for spans, and img tags. Allowed it to get links of the images. 
Although this is primarily meant for Twitter, the way I wrote it has allowed it to be inadvertentely used for all sites. 
Change is required inorder to make it Twitter specific.